requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
ctransformers==0.2.27
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-huggingface==0.1.2
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.7
langsmith==0.3.18
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.2.4
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
portalocker==2.10.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
qdrant-client==1.13.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
triton==3.2.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
zstandard==0.23.0
